Move SolarmanEmu class in a dedicated file
