+++ app/src/gen3plus/inverter_g3p.py
@@ -2,10 +2,12 @@ from asyncio import StreamReader, StreamWriter
 
 if __name__ == "app.src.gen3plus.inverter_g3p":
     from app.src.inverter_base import InverterBase
-    from app.src.gen3plus.solarman_v5 import SolarmanV5, SolarmanEmu
+    from app.src.gen3plus.solarman_v5 import SolarmanV5
+    from app.src.gen3plus.solarman_emu import SolarmanEmu
 else:  # pragma: no cover
     from inverter_base import InverterBase
-    from gen3plus.solarman_v5 import SolarmanV5, SolarmanEmu
+    from gen3plus.solarman_v5 import SolarmanV5
+    from gen3plus.solarman_emu import SolarmanEmu
 
 
 class InverterG3P(InverterBase):
